If PinMode is an enum class, then Camel has out of the box type converter to/from String for that guy, and therefore not really needed.
