+++ components/camel-raspberry/src/test/java/org/apache/camel/test/raspberry/output/DigitalOutput3Test.java
@@ -54,7 +54,7 @@ public class DigitalOutput3Test extends CamelTestSupport {
         return new RouteBuilder() {
             public void configure() {
                 from("timer://foo?repeatCount=2").id("rbpi-route").to("log:org.apache.camel.component.raspberry?showAll=true&multiline=true")
-                    .to("rbpi://pin?id=5&mode=output&state=LOW&action=TOGGLE").to("mock:result");
+                    .to("rbpi://pin?id=5&mode=DIGITAL_OUTPUT&state=LOW&action=TOGGLE").to("mock:result");
 
             }
         };
